Adding iverse_of in some relations
It's mandatory to have the inverse_of when validating the relation presence,
otherwise the relation model will not be valid, and the model
wont be valid in consequence.

This could be observed when adding a category to a new model,
it would not save, because the direct_category_items relation model
did not have the categorizable_id.

This way, it will follow the inverse relation and save
everybody, and we all.